chore: move ServerModule trait into fedimint-server
