+++ fedimint-dbtool/src/dump.rs
@@ -6,7 +6,7 @@ use erased_serde::Serialize;
 use fedimint_client::db::{ClientConfigKey, OperationLogKeyPrefix};
 use fedimint_client::module::init::ClientModuleInitRegistry;
 use fedimint_client::oplog::OperationLogEntry;
-use fedimint_core::config::{ClientConfig, CommonModuleInitRegistry, ServerModuleInitRegistry};
+use fedimint_core::config::{ClientConfig, CommonModuleInitRegistry};
 use fedimint_core::core::ModuleKind;
 use fedimint_core::db::{
     Database, DatabaseTransaction, DatabaseVersionKey, IDatabaseTransactionOpsCore,
@@ -19,6 +19,7 @@ use fedimint_rocksdb::RocksDbReadOnly;
 use fedimint_server::config::io::read_server_config;
 use fedimint_server::config::ServerConfig;
 use fedimint_server::consensus::db as ConsensusRange;
+use fedimint_server::core::{ServerModuleInitRegistry, ServerModuleInitRegistryExt};
 use fedimint_server::net::api::announcement::ApiAnnouncementPrefix;
 use futures::StreamExt;
 use ln_gateway::Gateway;
