+++ fedimint-dbtool/src/lib.rs
@@ -12,9 +12,7 @@ use anyhow::Result;
 use bytes::Bytes;
 use clap::{Parser, Subcommand};
 use fedimint_client::module::init::{ClientModuleInit, ClientModuleInitRegistry};
-use fedimint_core::config::ServerModuleInitRegistry;
 use fedimint_core::db::{IDatabaseTransactionOpsCore, IRawDatabaseExt};
-use fedimint_core::module::ServerModuleInit;
 use fedimint_core::util::handle_version_hash_command;
 use fedimint_ln_client::LightningClientInit;
 use fedimint_ln_server::LightningInit;
@@ -23,6 +21,7 @@ use fedimint_meta_client::MetaClientInit;
 use fedimint_meta_server::MetaInit;
 use fedimint_mint_client::MintClientInit;
 use fedimint_mint_server::MintInit;
+use fedimint_server::core::{ServerModuleInit, ServerModuleInitRegistry};
 use fedimint_wallet_client::WalletClientInit;
 use fedimint_wallet_server::WalletInit;
 use futures::StreamExt;
