+++ fedimint-recoverytool/src/main.rs
@@ -14,16 +14,17 @@ use clap::{ArgGroup, Parser, Subcommand};
 use fedimint_core::core::LEGACY_HARDCODED_INSTANCE_ID_WALLET;
 use fedimint_core::db::{Database, IDatabaseTransactionOpsCoreTyped};
 use fedimint_core::epoch::ConsensusItem;
+use fedimint_core::fedimint_build_code_version_env;
 use fedimint_core::module::registry::{ModuleDecoderRegistry, ModuleRegistry};
 use fedimint_core::module::CommonModuleInit;
 use fedimint_core::session_outcome::SignedSessionOutcome;
 use fedimint_core::transaction::Transaction;
 use fedimint_core::util::handle_version_hash_command;
-use fedimint_core::{fedimint_build_code_version_env, ServerModule};
 use fedimint_logging::TracingSetup;
 use fedimint_rocksdb::RocksDbReadOnly;
 use fedimint_server::config::io::read_server_config;
 use fedimint_server::consensus::db::SignedSessionOutcomePrefix;
+use fedimint_server::core::ServerModule;
 use fedimint_wallet_server::common::config::WalletConfig;
 use fedimint_wallet_server::common::keys::CompressedPublicKey;
 use fedimint_wallet_server::common::tweakable::Tweakable;
