+++ fedimint-server/src/consensus/engine.rs
@@ -17,7 +17,7 @@ use fedimint_core::endpoint_constants::AWAIT_SIGNED_SESSION_OUTCOME_ENDPOINT;
 use fedimint_core::epoch::ConsensusItem;
 use fedimint_core::fmt_utils::OptStacktrace;
 use fedimint_core::module::audit::Audit;
-use fedimint_core::module::registry::{ModuleDecoderRegistry, ServerModuleRegistry};
+use fedimint_core::module::registry::ModuleDecoderRegistry;
 use fedimint_core::module::{ApiRequestErased, SerdeModuleEncoding};
 use fedimint_core::net::peers::{DynP2PConnections, IP2PConnections};
 use fedimint_core::runtime::spawn;
@@ -28,6 +28,7 @@ use fedimint_core::task::{sleep, TaskGroup, TaskHandle};
 use fedimint_core::timing::TimeReporter;
 use fedimint_core::util::FmtCompact as _;
 use fedimint_core::{timing, NumPeers, NumPeersExt, PeerId};
+use fedimint_server_core::{ServerModuleRegistry, ServerModuleRegistryExt};
 use futures::StreamExt;
 use rand::Rng;
 use tokio::sync::watch;
