+++ fedimint-server/src/consensus/mod.rs
@@ -16,16 +16,15 @@ use async_channel::Sender;
 use db::get_global_database_migrations;
 use fedimint_api_client::api::net::Connector;
 use fedimint_api_client::api::{DynGlobalApi, P2PConnectionStatus};
-use fedimint_core::config::ServerModuleInitRegistry;
 use fedimint_core::core::{ModuleInstanceId, ModuleKind};
 use fedimint_core::db::{apply_migrations, apply_migrations_server, Database};
 use fedimint_core::envs::is_running_in_test_env;
 use fedimint_core::epoch::ConsensusItem;
 use fedimint_core::module::registry::ModuleRegistry;
-use fedimint_core::server::DynServerModule;
 use fedimint_core::task::TaskGroup;
 use fedimint_core::NumPeers;
 use fedimint_logging::{LOG_CONSENSUS, LOG_CORE};
+use fedimint_server_core::{DynServerModule, ServerModuleInitRegistry};
 use jsonrpsee::server::ServerHandle;
 use tokio::sync::watch;
 use tracing::{info, warn};
