+++ fedimint-testing-core/Cargo.toml
@@ -21,6 +21,7 @@ fedimint-client = { workspace = true }
 fedimint-core = { workspace = true }
 fedimint-logging = { workspace = true }
 fedimint-rocksdb = { workspace = true }
+fedimint-server = { workspace = true }
 futures = { workspace = true }
 rand = { workspace = true }
 tempfile = "3.15.0"
