+++ modules/fedimint-dummy-server/Cargo.toml
@@ -21,6 +21,7 @@ async-trait = { workspace = true }
 erased-serde = { workspace = true }
 fedimint-core = { workspace = true }
 fedimint-dummy-common = { workspace = true }
+fedimint-server-core = { workspace = true }
 futures = { workspace = true }
 serde = { workspace = true }
 strum = { workspace = true }
