+++ modules/fedimint-empty-server/Cargo.toml
@@ -21,6 +21,7 @@ async-trait = { workspace = true }
 erased-serde = { workspace = true }
 fedimint-core = { workspace = true }
 fedimint-empty-common = { workspace = true }
+fedimint-server-core = { workspace = true }
 futures = { workspace = true }
 serde = { workspace = true }
 strum = { workspace = true }
