+++ modules/fedimint-ln-tests/Cargo.toml
@@ -24,6 +24,7 @@ fedimint-ln-client = { workspace = true }
 fedimint-ln-common = { workspace = true }
 fedimint-ln-server = { workspace = true }
 fedimint-logging = { workspace = true }
+fedimint-server = { workspace = true }
 fedimint-testing = { workspace = true }
 futures = { workspace = true }
 lightning-invoice = { workspace = true }
