+++ modules/fedimint-mint-server/src/lib.rs
@@ -22,14 +22,13 @@ use fedimint_core::db::{
 use fedimint_core::module::audit::Audit;
 use fedimint_core::module::{
     api_endpoint, ApiEndpoint, ApiError, ApiVersion, CoreConsensusVersion, InputMeta,
-    ModuleConsensusVersion, ModuleInit, PeerHandle, ServerModuleInit, ServerModuleInitArgs,
-    SupportedModuleApiVersions, TransactionItemAmount, CORE_CONSENSUS_VERSION,
+    ModuleConsensusVersion, ModuleInit, PeerHandle, SupportedModuleApiVersions,
+    TransactionItemAmount, CORE_CONSENSUS_VERSION,
 };
-use fedimint_core::server::DynServerModule;
 use fedimint_core::util::BoxFuture;
 use fedimint_core::{
     apply, async_trait_maybe_send, push_db_key_items, push_db_pair_items, secp256k1, Amount,
-    NumPeersExt, OutPoint, PeerId, ServerModule, Tiered, TieredMulti,
+    NumPeersExt, OutPoint, PeerId, Tiered, TieredMulti,
 };
 use fedimint_logging::LOG_MODULE_MINT;
 pub use fedimint_mint_common as common;
@@ -46,6 +45,9 @@ use fedimint_mint_common::{
 };
 use fedimint_server::config::distributedgen::{eval_poly_g2, PeerHandleOps};
 use fedimint_server::consensus::db::{MigrationContextExt, TypedModuleHistoryItem};
+use fedimint_server::core::{
+    DynServerModule, ServerModule, ServerModuleInit, ServerModuleInitArgs,
+};
 use futures::StreamExt;
 use itertools::Itertools;
 use metrics::{
@@ -809,10 +811,11 @@ mod test {
     use fedimint_core::db::mem_impl::MemDatabase;
     use fedimint_core::db::Database;
     use fedimint_core::module::registry::ModuleRegistry;
-    use fedimint_core::module::{ModuleConsensusVersion, ServerModuleInit};
-    use fedimint_core::{secp256k1, Amount, PeerId, ServerModule};
+    use fedimint_core::module::ModuleConsensusVersion;
+    use fedimint_core::{secp256k1, Amount, PeerId};
     use fedimint_mint_common::config::FeeConsensus;
     use fedimint_mint_common::{MintInput, Nonce, Note};
+    use fedimint_server::core::{ServerModule, ServerModuleInit};
     use tbs::blind_message;
 
     use crate::common::config::MintGenParamsConsensus;
