+++ modules/fedimint-unknown-server/Cargo.toml
@@ -20,6 +20,7 @@ anyhow = { workspace = true }
 async-trait = { workspace = true }
 erased-serde = { workspace = true }
 fedimint-core = { workspace = true }
+fedimint-server-core = { workspace = true }
 fedimint-unknown-common = { workspace = true }
 strum = { workspace = true }
 strum_macros = { workspace = true }
